+++ tests/greedysearch_test.cpp
@@ -6,7 +6,7 @@
 using namespace std;
 
 
-Node* create_node(unsigned int id, const vector<double>& coords) {
+Node* create_node(unsigned int id, const vector<float>& coords) {
     Node* node = new Node();
     node->id = id;
     node->coords = coords;
